take first capability from allowed as default
Before https://github.com/canonical/subiquity/pull/1774 the order of the capabilities in `allowed` was arbitrary, resulting in https://bugs.launchpad.net/ubuntu-desktop-installer/+bug/2033093 and other confusing behaviour. The subiquity PR puts the capability that should be considered the default first, so that's what the client side code should do to.