fix: DBus SetPassword method expects a crypted pass
The method org.freedesktop.Accounts.User.SetPassword expects
a crypted password as its first parameter. Unfortunately, the
current code is sending the password in plain text. That plain
text password is being stored as-is in the 'shadow' file, and
that prevents ubuntu-core-desktop-init to work on core desktop,
because the created user has an invalid password field.

This patch fixes it.