feat: show source type even if only 1 option
When we started the move to a single 'default' installation the install
type radio button was redundant and this section was removed.
Subsequently, it was decided to revert to two installation options
since for 23.10 we are not able to deliver the wider redesign. So,
this change is to always show the installation types as reported by
subiquity (even if there is only one).