Add correct end page for the Core Desktop installer
The current Ubuntu-core-desktop-init program used for the
installer shows, at the end, the same page than the live
installer: one where the user can choose between “reboot” or
“continue testing”. This is invalid for Core Desktop because
it’s not a live session, so choosing “continue testing” leaves
a graphical session without desktop.

This patch adds a new command-line parameter, core-install,
that, when set to 1, will change the last page to one that
only allows to reboot, and shows a more fitting message.