Add flavo(u)r data in ProductService
It can be useful to have available the flavor data passed during
the call to runInstallerApp (for example, to change the page
appearance for specific Ubuntu flavors, like Ubuntu Core Desktop,
based on the id field).

This MR makes that data available in the ProductService class,
thus allowing to retrieve it anywhere.