ci: Automate rebase to support multiple snaps
Launchpad lacks a pre-build hook for snap builders; so without changes
we cannot build multiple snaps from one branch. Using a GitHub action
we could reconfigure the repo and then invoke a remote build. But we'd
lose Launchpad's audit trail which is needed for a seeded snap.
Alternatively, we could continue using branches-per-snap [1][2] but this
means silo'ing ci artefacts to branches.

This commit triggers on workflow `main` changes to resolve this. The
workflow:

1. rebases `ci/*` branches on `main`.
2. repositions artefacts for the Launchpad builder jobs.
3. amends the HEAD commit to avoid piling up automated commits.

Note: The workflow does not sign step 3's amended.

References:

[1] https://launchpad.net/~ubuntu-desktop/ubuntu-desktop-provision/+snap/ubuntu-desktop-bootstrap
[2] https://launchpad.net/~ubuntu-desktop/ubuntu-desktop-provision/+snap/ubuntu-desktop-init