ci: Move snapcraft build files into ci folder
Snapcraft.yamls and artefacts were in the /snap/* folder. This meant
that running `snapcraft` gave an error. It was also brittle when running
local builds due to individual files within snap/ needing to be move
around (rather than simply being able to remove the folder and copy a
fresh one from somewhere else).

This restructure avoids the warning and provides a straightforward way
to prepare for a snap build:

```bash
$ ./ci/snap-build-setup.sh bootstrap # or
$ ./ci/snap-build-setup.sh init
```