ci(init,bootstrap): Improve resiliency of snapcraft.yamls
Building locally is brittle due to artefacts polluting the build
container (caused by `source: .`). This doesn't present in ci because
the build container is fresh.

The flutter-git part, now checks if a previous clone exists to avoid
a failure if the snapcraft build is run twice. Additionally,
CONTRIBUTING.md highlights the difficulties with local builds and
provides some guidance on workarounds.