snapcraft: set PATH/PYTHON/PYTHONPATH _ORIG
Ensure subiquity-server and ubuntu-desktop-bootstrap apps have
PATH_ORIG, PYTHON_ORIG, and PYTHONPATH_ORIG environment variables.
Lacking these causes usage of subiquity's system_scripts to fail to find
the python outside of the snap. This will result in blocking installs
since suiquity relies on cloud-init scripts to validate user data.

A special note on PATH_ORIG: For some unknown reason - potentially a
snapcraft bug - when setting environment variables in the subiquity-server
environment, the PATH variable has already been modified to include
paths with a SNAP prefix. This means we can't set PATH_ORIG to PATH
like we do in subiquity. Instead we use the documented default of PATH
for classic confinement snaps in [0] as the value for PATH_ORIG to ensure
we get only paths to outside of the snap.

[0] https://snapcraft.io/docs/environment-variables#heading--path