fix: correctly update button states on NetworkPage
Instead of showing both a 'Connect' and a 'Next' button (where one of
those is hidden), this ensures only one of the widgets is present in the
widget tree and its corresponding state gets reset when the buttons are
exchanged. This fixes a bug where the 'Connect' button gets stuck in a
loading state and also stops the button(s) from moving around.