refactor: break up `AutoinstallModel`
`AutoinstallModel` now maintains the overall state:
- none: interactive installation
- direct: directly fetch autoinstall.yaml from URL or local file path
- landscape: future work - fetch autoinstall.yaml from landscape server

It also provides methods to retrieve the current
autoinstall.yaml (in the liveISO root or the .subiquity dir) and restart
subiquity + the UI.

`AutoinstallDirectModel` will hold the UI state for the text field and
file picker used to directly fetch an autoinstall.yaml from a URL or
local path. The methods for those actions will be moved into the new
service in the next refactoring step.