fix(confirm): restore modified disks (#1018)
In #969 we've changed the behavior of the `confirmModel` to provide a
list of *all* disks instead of only those that have been modified in
order to include unchanged partitions in the table on the confirmation
page. However, the 'Installation disk' entry on that page still needs
access to a list of modified disks - this PR (re-)adds this property.

Fix #1012
UDENG-6572